pyproject.toml
[project]
name = "branthebuilder"
dynamic = ["version", "description"]
authors = [{name = "Endre Márk Borza", email = "endremborza@gmail.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer",
    "flit",
    "jupyter",
    "black",
    "pytest",
    "pytest-cov",
    "codecov",
    "toml",
    "flake8",
    "isort",
    "cookiecutter",
    "pyyaml"
]

[project.optional-dependencies]
test = []
doc = [
    "sphinx",
    "graphviz",
    "pandoc",
    "sphinx-automodapi",
    "sphinx-rtd-theme",
    "myst-parser",
    "pygments",
]

[project.urls]
Homepage = "https://github.com/endremborza/branthebuilder"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.branb]
line-length = 88


[project.scripts]
branb = "branthebuilder:app"
